pyproject.toml
[project]
name = "aiogram_bot_template"
version = "1.0"
description = "Aiogram 3.x bot template using PostgreSQL (asyncpg) with SQLAlchemy + alembic"
authors = ["wakaree <nullmatawasoradesu@gmail.com>"]
readme = "README.md"
license = "MIT"
requires-python = ">=3.12,<3.13"
dependencies = [
    "aiogram~=3.20.0",
    "aiogram-contrib>=1.1.4",
    "aiogram-i18n~=1.4",
    "aiohttp~=3.11.18",
    "alembic~=1.14.1",
    "asyncpg~=0.30.0",
    "fastapi>=0.115.12",
    "fluent-runtime~=0.4.0",
    "greenlet>=3.2.2",
    "msgspec~=0.18.6",
    "pydantic~=2.10.6",
    "pydantic-settings[yaml]~=2.9.1",
    "redis~=5.2.1",
    "sqlalchemy~=2.0.41",
    "uvicorn>=0.34.2",
]

[project.urls]
Repository = "https://github.com/wakaree/aiogram_bot_template.git"

[tool.uv]
dev-dependencies = [
    "ftl-extract>=0.8.0",
    "mypy>=1.14.1",
    "ruff>=0.8.6",
]

[tool.black]
line-length = 99
exclude = "\\.?venv|\\.?tests"

[tool.ruff]
target-version = "py38"
line-length = 99
lint.select = [
    "C",
    "DTZ",
    "E",
    "F",
    "I",
    "ICN",
    "N",
    "PLC",
    "PLE",
    "Q",
    "T",
    "W",
    "YTT",
]
lint.ignore = ["N805"]
exclude = [
    ".venv",
    ".idea",
]
[tool.mypy]
plugins = [
    "sqlalchemy.ext.mypy.plugin",
    "pydantic.mypy"
]
exclude = [
    "venv",
    ".venv",
    ".idea",
    ".tests",
]
warn_unused_configs = true
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
warn_unused_ignores = true
warn_return_any = true
no_implicit_reexport = true
strict_equality = true
extra_checks = true

[[tool.mypy.overrides]]
module = ["app.telegram.handlers.*"]
strict_optional = false
warn_return_any = false
disable_error_code = ["union-attr"]
